refactor: change project package
