--- app/src/main/java/com/example/judgement/api/MyPreference.kt
@@ -1,17 +1,0 @@
-package com.example.judgement.api
-
-import android.app.Application
-
-class MyPreference : Application()
-{
-    companion object
-    {
-        lateinit var prefs: PreferenceUtil
-    }
-
-    override fun onCreate()
-    {
-        prefs = PreferenceUtil(applicationContext)
-        super.onCreate()
-    }
-}
