--- app/src/main/java/com/example/judgement/ui/navigation/scrap/ScrapManager.kt
@@ -1,5 +1,0 @@
-package com.example.judgement.ui.navigation.scrap
-
-data class ScrapManager(
-    val scrapRvAdapters: MutableMap<String, ScrapRvAdapter> = mutableMapOf(), // 어댑터
-)
