+++ app/src/main/java/com/example/judgement/util/MyPreference.kt
@@ -1,0 +1,17 @@
+package com.example.judgement.util
+
+import android.app.Application
+
+class MyPreference : Application()
+{
+    companion object
+    {
+        lateinit var prefs: PreferenceUtil
+    }
+
+    override fun onCreate()
+    {
+        prefs = PreferenceUtil(applicationContext)
+        super.onCreate()
+    }
+}
