+++ app/src/main/java/com/example/judgement/util/PreferenceUtil.kt
@@ -1,0 +1,20 @@
+package com.example.judgement.util
+import android.content.Context
+import android.content.SharedPreferences
+
+
+class PreferenceUtil (context: Context)
+{
+    private val prefs: SharedPreferences = context.getSharedPreferences("other2", 0)
+
+    fun getString(key: String, defValue: String): String
+    {
+        return prefs.getString(key, defValue).toString()
+    }
+
+    fun setString(key: String, str: String)
+    {
+        prefs.edit().putString(key, str).apply()
+    }
+
+}
