+++ app/src/main/res/layout/activity_detail_result.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="activity"
-            type="com.example.judgement.ui.law.DetailResultActivity" />
+            type="com.example.judgement.view.detail_result.DetailResultActivity" />
 
     </data>
 
@@ -15,7 +15,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".ui.law.DetailResultActivity">
+        tools:context=".view.detail_result.DetailResultActivity">
 
         <!-- top -->
         <LinearLayout
