+++ app/src/main/res/layout/activity_main.xml
@@ -6,13 +6,13 @@
     <data>
         <variable
             name="activity"
-            type="com.example.judgement.ui.navigation.MainActivity" />
+            type="com.example.judgement.view.main.MainActivity" />
 
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.navigation.MainActivity">
+        tools:context=".view.main.MainActivity">
 
         <FrameLayout
             android:id="@+id/frame_layout_main"
