+++ app/src/main/res/layout/activity_news.xml
@@ -7,13 +7,13 @@
     <data>
         <variable
             name="activity_news"
-            type="com.example.judgement.ui.navigation.home.NewsActivity" />
+            type="com.example.judgement.view.main.home.news.NewsActivity" />
     </data>
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.navigation.home.NewsActivity">
+        tools:context=".view.main.home.news.NewsActivity">
 
         <WebView
             android:id="@+id/web_view_news"
