+++ app/src/main/res/layout/activity_search_result.xml
@@ -6,14 +6,14 @@
     <data>
         <variable
             name="activity_search_result"
-            type="com.example.judgement.ui.law.search_result.SearchResultActivity" />
+            type="com.example.judgement.view.search_result.SearchResultActivity" />
     </data>
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".ui.law.search_result.SearchResultActivity">
+        tools:context=".view.search_result.SearchResultActivity">
 
         <LinearLayout
             style="@style/TextPadding"
