+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="activity"
-            type="com.example.judgement.ui.signup.SignUpActivity" />
+            type="com.example.judgement.view.signup.SignUpActivity" />
 
     </data>
