+++ app/src/main/res/layout/fragment_category.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="fragment_category"
-            type="com.example.judgement.ui.navigation.category.CategoryFragment" />
+            type="com.example.judgement.view.main.category.CategoryFragment" />
     </data>
 
     <LinearLayout
@@ -13,7 +13,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".ui.navigation.category.CategoryFragment">
+        tools:context=".view.main.category.CategoryFragment">
 
         <LinearLayout
             style="@style/TextPadding"
