+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="fragment_home"
-            type="com.example.judgement.ui.navigation.home.HomeFragment" />
+            type="com.example.judgement.view.main.home.HomeFragment" />
     </data>
 
     <androidx.core.widget.NestedScrollView
@@ -16,7 +16,7 @@
         android:layout_height="match_parent"
         android:overScrollMode="never"
         app:layout_scrollFlags="scroll"
-        tools:context=".ui.navigation.home.HomeFragment">
+        tools:context=".view.main.home.HomeFragment">
 
         <LinearLayout
             android:layout_width="match_parent"
