+++ app/src/main/res/layout/fragment_scrap.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="activity"
-            type="com.example.judgement.ui.navigation.scrap.ScrapFragment" />
+            type="com.example.judgement.view.main.scrap.ScrapFragment" />
 
     </data>
     <LinearLayout
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".ui.navigation.scrap.ScrapFragment">
+        tools:context=".view.main.scrap.ScrapFragment">
 
         <!-- top -->
         <TextView
