+++ app/src/main/res/layout/fragment_user.xml
@@ -6,14 +6,14 @@
     <data>
         <variable
             name="activity"
-            type="com.example.judgement.ui.navigation.user.UserFragment" />
+            type="com.example.judgement.view.main.user.UserFragment" />
 
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/fragment_user"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.navigation.user.UserFragment">
+        tools:context=".view.main.user.UserFragment">
 
         <!-- top -->
         <TextView
