+++ app/src/main/res/layout/fragment_user_info_edit.xml
@@ -6,14 +6,14 @@
     <data>
         <variable
             name="activity"
-            type="com.example.judgement.ui.navigation.user.UserInfoEditFragment" />
+            type="com.example.judgement.view.main.user.UserInfoEditFragment" />
 
     </data>
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".ui.navigation.user.UserInfoEditFragment">
+        tools:context=".view.main.user.UserInfoEditFragment">
 
         <!-- top -->
         <androidx.constraintlayout.widget.ConstraintLayout
